[Platform]: Moved Data download to common components folder (#405)
