+++ apps/platform/src/theme.js
@@ -1,5 +1,6 @@
 import { lighten, darken } from "polished";
 import config from "./config";
+import { grey } from "@mui/material/colors";
 
 const PRIMARY = config.profile.primaryColor;
 const SECONDARY = "#ff6350";
@@ -58,6 +59,15 @@ const theme = {
           textTransform: "none",
           color: "#5A5F5F",
           borderColor: "rgb(196,196,196)",
+        },
+      },
+    },
+    MuiMenu: {
+      styleOverrides: {
+        paper: {
+          boxShadow: "none",
+          border: `1px solid ${grey[400]}`,
+          marginTop: -1,
         },
       },
     },
